refact for lint
